requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.7
python-dotenv==0.19.0
requests==2.31.0 